+++ non-standard-plugins.js
@@ -4,6 +4,7 @@ module.exports = function shopifyNonStandardPlugins(options = {}) {
   const plugins = [
     require.resolve('@babel/plugin-proposal-class-properties'),
     require.resolve('@babel/plugin-syntax-dynamic-import'),
+    require.resolve('@babel/plugin-proposal-optional-catch-binding'),
   ];
 
   if (inlineEnv) {